refactor(layout): moved layout to dedicated folder, utilized logout page
