+++ src/component/layout/Layout.tsx
@@ -1,11 +1,9 @@
-import AccountCircle from '@mui/icons-material/AccountCircle';
 import {
     AppBar,
     Box,
     Container,
     CssBaseline,
     Drawer,
-    IconButton,
     List,
     ListItemButton,
     ListItemText,
@@ -13,11 +11,12 @@ import {
     Typography
 } from '@mui/material';
 import { Link, Outlet } from 'react-router-dom';
-import Privilege from '../model/member/Privilege.ts';
-import CardWrapperComponent from '../page/member/CardWrapperComponent.tsx';
-import WEB_ENDPOINTS from '../util/endpoint/WebEndpoint.ts';
-import ButtonCollapseList from './ButtonCollapseList.tsx';
-import EnableOnRole from './EnableOnRole.tsx';
+import Privilege from '../../model/member/Privilege.ts';
+import CardWrapperComponent from '../../page/member/CardWrapperComponent.tsx';
+import WEB_ENDPOINTS from '../../util/endpoint/WebEndpoint.ts';
+import ButtonCollapseList from '../ButtonCollapseList.tsx';
+import EnableOnRole from '../EnableOnRole.tsx';
+import AccountButton from './AccountButton.tsx';
 
 
 const drawerWidth = 240;
@@ -42,16 +41,7 @@ const Layout = () => {
                             PassGo
                         </Link>
                     </Typography>
-                    <IconButton
-                        size="large"
-                        aria-label="account of current user"
-                        aria-controls="menu-appbar"
-                        aria-haspopup="true"
-                        color="inherit"
-                        component={Link} to='/members/me'
-                    >
-                        <AccountCircle />
-                    </IconButton>
+                    <AccountButton />
                 </Toolbar>
             </AppBar>
